fix: client view
